fixed: #141
new use of the SearchWaitDialog. it displays now the loaded results
in a progress bar. the user can cancel the search.
The SearchResultTree is overwritten to not sort the results, as
it would be fill the MetaObjectNodes. The Workbench manages his own
table sorting.
The new "init" method in BaseEntity allows to prefill some stuff
after bean construction. in example: the leuchten object in standort