issue #160
implementation of the QuerySearchResultsWindowSearch. This WindowSearch
wraps a QuerySearchResultsActionPanel, and implements two actions
One for the csv export of the search results, and the other doing a full
search and passing the results to the workbench.

The BelisWindowSearch class has be moved to BelisEntityWindowSearch.
Now BelisWindowSearch is the interface that have to be used as ServiceProvider
for the integration of a WindowSearch in Belis.